examples/sftp-automated-workflows-agentcore/agent-source-code/fraud_detection_agent/requirements.txt
bedrock-agentcore
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
httpx>=0.27.0
mcp>=1.0.0
